poprawki zwiazane z integracja z timelinem
git-svn-id: https://localhost:8443/svn/hm-mde-1@121 fca3399c-2208-499e-aaa1-2cf7fb07c25e

+++ branches/chart/ChartService.cpp
@@ -74,7 +74,7 @@ ChartService::ChartService()
   length(0.0)
 {
     widget = new ChartWidget(this);
-	length=widget->getViewer()->getData()->getRNumber()/widget->getViewer()->getData()->getFPS();
+	length=(double)widget->getViewer()->getData()->getRNumber()/(double)widget->getViewer()->getData()->getFPS() ;
 }
